test(desktop): 添付 test を分割し、案内 CSS を composer hint の隣へ移動 (#965)
- CI の oversized-files が 1000 行超の新規 file として useDesktopShellActions.test.tsx と shell-phase1-part1.css を検出したため、baseline を更新せず分割
- renderActionsHook などの共通ハーネスを shell/testSupport/renderShellActions.tsx へ抽出し、添付の handler test を useDesktopShellActions.attachments.test.tsx へ移動
- .composer-attachment-formats を shell-phase1-part4.css の composer hint の隣へ移動
- before / after screenshot を docs/progress/assets/issue-965 に追加し、記録を更新

Refs #965
